fix(S5): int заменен на long long + CG

+++ miheev.egor/S5/IOFunctions.cpp
@@ -2,7 +2,7 @@
 
 void miheev::readTree(std::istream& in, tree_t& container)
 {
-  int tempKey = 0;
+  long long tempKey = 0;
   std::string tempValue = "";
   while (!in.eof())
   {
